+++ src/icemac/addressbook/browser/addressbook/crud.py
@@ -1,9 +1,8 @@
 # -*- coding: latin-1 -*-
 # Copyright (c) 2008-2013 Michael Howitz
 # See also LICENSE.txt
-# $Id$
-
-from icemac.addressbook.i18n import MessageFactory as _
+from icemac.addressbook.i18n import _
+import icemac.addressbook.addressbook
 import icemac.addressbook.browser.base
 import icemac.addressbook.browser.interfaces
 import icemac.addressbook.browser.metadata
@@ -11,6 +10,14 @@ import icemac.addressbook.interfaces
 import z3c.form.button
 import zope.interface
 import zope.size.interfaces
+
+
+class AddForm(icemac.addressbook.browser.base.BaseAddForm):
+
+    label = _(u'Add new address book')
+    interface = icemac.addressbook.interfaces.IAddressBook
+    class_ = icemac.addressbook.addressbook.AddressBook
+    next_url = 'object'
 
 
 class EditForm(icemac.addressbook.browser.base.GroupEditForm):